Refactor Unicode case-folding expansion
